Splitting large file Textures into multiple files
