+++ Sources/Laka/Subcommands/Images/Textures/TextureMonsters.swift
@@ -1,0 +1,177 @@
+//
+//  File.swift
+//  
+//
+//  Created by Alexander Cyon on 2021-10-11.
+//
+
+import Foundation
+
+// MARK: - Monsters
+// MARK: -
+extension Laka.Textures {
+    private var monsters: [String] { [
+        "avwpike.def",
+        "avwpikx0.def",
+        "avwlcrs.def",
+        "avwhcrs.def",
+        "avwgrif.def",
+        "avwgrix0.def",
+        "avwswrd0.def",
+        "avwswrx0.def",
+        "avwmonk.def",
+        "avwmonx0.def",
+        "avwcvlr0.def",
+        "avwcvlx0.def",
+        "avwangl.def",
+        "avwarch.def",
+        "avwcent0.def",
+        "avwcenx0.def",
+        "avwdwrf0.def",
+        "avwdwrx0.def",
+        "avwelfw0.def",
+        "avwelfx0.def",
+        "avwpega0.def",
+        "avwpegx0.def",
+        "avwtree0.def",
+        "avwtrex0.def",
+        "avwunic0.def",
+        "avwunix0.def",
+        "avwdrag0.def",
+        "avwdrax0.def",
+        "avwgrem0.def",
+        "avwgrex0.def",
+        "avwgarg0.def",
+        "avwgarx0.def",
+        "avwgolm0.def",
+        "avwgolx0.def",
+        "avwmage0.def",
+        "avwmagx0.def",
+        "avwgeni0.def",
+        "avwgenx0.def",
+        "avwnaga0.def",
+        "avwnagx0.def",
+        "avwtitn0.def",
+        "avwtitx0.def",
+        "avwimp0.def",
+        "avwimpx0.def",
+        "avwgog0.def",
+        "avwgogx0.def",
+        "avwhoun0.def",
+        "avwhoux0.def",
+        "avwdemn0.def",
+        "avwdemx0.def",
+        "avwpitf0.def",
+        "avwpitx0.def",
+        "avwefre0.def",
+        "avwefrx0.def",
+        "avwdevl0.def",
+        "avwdevx0.def",
+        "avwskel0.def",
+        "avwskex0.def",
+        "avwzomb0.def",
+        "avwzomx0.def",
+        "avwwigh.def",
+        "avwwigx0.def",
+        "avwvamp0.def",
+        "avwvamx0.def",
+        "avwlich0.def",
+        "avwlicx0.def",
+        "avwbkni0.def",
+        "avwbknx0.def",
+        "avwbone0.def",
+        "avwbonx0.def",
+        "avwtrog0.def",
+        "avwinfr.def",
+        "avwharp0.def",
+        "avwharx0.def",
+        "avwbehl0.def",
+        "avwbehx0.def",
+        "avwmeds.def",
+        "avwmedx0.def",
+        "avwmino.def",
+        "avwminx0.def",
+        "avwmant0.def",
+        "avwmanx0.def",
+        "avwrdrg.def",
+        "avwddrx0.def",
+        "avwgobl0.def",
+        "avwgobx0.def",
+        "avwwolf0.def",
+        "avwwolx0.def",
+        "avworc0.def",
+        "avworcx0.def",
+        "avwogre0.def",
+        "avwogrx0.def",
+        "avwroc0.def",
+        "avwrocx0.def",
+        "avwcycl0.def",
+        "avwcycx0.def",
+        "avwbhmt0.def",
+        "avwbhmx0.def",
+        "avwgnll0.def",
+        "avwgnlx0.def",
+        "avwlizr.def",
+        "avwlizx0.def",
+        "avwdfly.def",
+        "avwdfir.def",
+        "avwbasl.def",
+        "avwgbas.def",
+        "avwgorg.def",
+        "avwgorx0.def",
+        "avwwyvr.def",
+        "avwwyvx0.def",
+        "avwhydr.def",
+        "avwhydx0.def",
+        "avwpixie.def",
+        "avwsprit.def",
+        "avwelmw0.def",
+        "avwicee.def",
+        "avwelme0.def",
+        "avwstone.def",
+        "avwelma0.def",
+        "avwstorm.def",
+        "avwelmf0.def",
+        "avwnrg.def",
+        "avwpsye.def",
+        "avwmagel.def",
+        "avwfbird.def",
+        "avwphx.def",
+        "avwglmg0.def",
+        "avwglmd0.def",
+        "avwazure.def",
+        "avwcdrg.def",
+        "avwfdrg.def",
+        "avwrust.def",
+        "avwench.def",
+        "avwsharp.def",
+        "avwhalf.def",
+        "avwpeas.def",
+        "avwboar.def",
+        "avwmumy.def",
+        "avwnomd.def",
+        "avwrog.def",
+        "avwtrll.def",
+        "avwmrnd0.def",
+        "avwmon1.def",
+        "avwmon2.def",
+        "avwmon3.def",
+        "avwmon4.def",
+        "avwmon5.def",
+        "avwmon6.def",
+        "avwmon7.def"
+    ]
+    }
+    
+    func exportMonsters() throws {
+        let monsterFiles: [ImageExport] = monsters.map { defFileName in
+            ImageExport(defFileName: defFileName, nameFromFrameIndex: { _ in defFileName })
+        }
+        try generateTexture(
+            name: "monsters",
+            list: monsterFiles,
+            maxImageCountPerDefFile: 1
+        )
+    }
+    
+}
