+++ Sources/Laka/Subcommands/Images/Textures/TextureVisitable.swift
@@ -1,0 +1,197 @@
+//
+//  File.swift
+//  
+//
+//  Created by Alexander Cyon on 2021-10-11.
+//
+
+import Foundation
+
+extension Laka.Textures {
+    
+    private var visitable: [String] {
+        [
+        "avxsndg0.def",
+        "avxsnlv0.def",
+        "avxsnds0.def",
+        "avxsnsw0.def",
+        "avmwwhl0.def",
+        "avxsnsn0.def",
+        "avsclvd0.def",
+        "avsfntn0.def",
+        "avsring0.def",
+        "avtmyst0.def",
+        "avxfyth0.def",
+        "avxgyne0.def",
+        "avxhild0.def",
+        "avxoblg.def",
+        "avxpost0.def",
+        "avxstbl0.def",
+        "avxtrek0.def",
+        "avxwelg0.def",
+        "avadlic0.def",
+        "avawrek0.def",
+        "avawre20.def",
+        "avsbuoy0.def",
+        "avxmapw0.def",
+        "avxmerm0.def",
+        "avxsirn0.def",
+        "avxwhrl0.def",
+        "avsclvg0.def",
+        "avxhilg0.def",
+        "avxoblw.def",
+        "avxobly.def",
+        "avtwagn0.def",
+        "avxmags0.def",
+        "avxoblo.def",
+        "avxpstr0.def",
+        "avxskds0.def",
+        "avxwtrh0.def",
+        "avxgyds0.def",
+        "avxoblk.def",
+        "avxosis0.def",
+        "avxprmd0.def",
+        "avsclvs0.def",
+        "avxoblb.def",
+        "avmlean0.def",
+        "avmwmsn0.def",
+        "avmwwsn0.def",
+        "avxgysn0.def",
+        "avxoblp.def",
+        "avxpssn.def",
+        "avxreds0.def",
+        "avxwlsn0.def",
+        "avxmapu0.def",
+        "avxpllr0.def",
+        "avgsieg0.def",
+        "avgrefg0.def",
+        "avmwndd0.def",
+        "avsarna0.def",
+        "avsaxis0.def",
+        "avsgrdn0.def",
+        "avsgzbo0.def",
+        "avsidol0.def",
+        "avslibr0.def",
+        "avsmarl.def",
+        "avsmerc0.def",
+        "avsschm0.def",
+        "avsuniv0.def",
+        "avstmpl0.def",
+        "avsutop0.def",
+        "avswar20.def",
+        "avxaltar.def",
+        "avxbnk10.def",
+        "avxbnk20.def",
+        "avxbnk30.def",
+        "avxbnk40.def",
+        "avxbnk50.def",
+        "avxbnk60.def",
+        "avxbnk70.def",
+        "avxcovr0.def",
+        "avxdent.def",
+        "avxdend0.def",
+        "avxeyem0.def",
+        "avxhutm0.def",
+        "avxmaps0.def",
+        "avxmktb0.def",
+        "avxmktt0.def",
+        "avxmn1b0.def",
+        "avxmn1r0.def",
+        "avxmn1y0.def",
+        "avxmn4i0.def",
+        "avxmn5i0.def",
+        "avxmn6i0.def",
+        "avxmn7i0.def",
+        "avxmn8i0.def",
+        "avxmx1b0.def",
+        "avxmx1r0.def",
+        "avxmx1y0.def",
+        "avxmn4o0.def",
+        "avxmn5o0.def",
+        "avxmn6o0.def",
+        "avxmn7o0.def",
+        "avxmn8o0.def",
+        "avxmn2g0.def",
+        "avxmn2o0.def",
+        "avxmn2p0.def",
+        "avxmn4b0.def",
+        "avxmn5b0.def",
+        "avxmn6b0.def",
+        "avxmn7b0.def",
+        "avxmn8b0.def",
+        "avxbor00.def",
+        "avxbor10.def",
+        "avxbor20.def",
+        "avxbor30.def",
+        "avxbor40.def",
+        "avxbor50.def",
+        "avxbor60.def",
+        "avxbor70.def",
+        "avxbgt00.def",
+        "avxbgt10.def",
+        "avxbgt20.def",
+        "avxbgt30.def",
+        "avxbgt40.def",
+        "avxbgt50.def",
+        "avxbgt60.def",
+        "avxbgt70.def",
+        "avxfgld.def",
+        "avxkey00.def",
+        "avxkey10.def",
+        "avxkey20.def",
+        "avxkey30.def",
+        "avxkey40.def",
+        "avxkey50.def",
+        "avxkey60.def",
+        "avxkey70.def",
+        "avxredw.def",
+        "avxrlly0.def",
+        "avxsanc0.def",
+        "avxtomb0.def",
+        "avxtvrn0.def",
+        "avxwelr0.def",
+        "avxboat0.def",
+        "avxboat1.def",
+        "avxboat2.def",
+        "avxbttl0.def",
+        "avzevnt0.def",
+        "avxlths0.def",
+        "avxprsn0.def",
+        "avxbor80.def",
+        "avxschl0.def",
+        "avxseeb0.def",
+        "avxseer0.def",
+        "avxseey0.def",
+        "avxshyd0.def",
+        "avxl1sh0.def",
+        "avxl2sh0.def",
+        "avxl3sh0.def",
+        "avtcave.def",
+        "avswtch0.def",
+        "avcgar10.def",
+        "avcgar20.def",
+        "avcvgr.def",
+        "avcvgarm.def",
+        "avaflot0.def",
+        "avasurv0.def",
+        "avxccht0.def",
+        "avxcflv0.def",
+        "avxcfds0.def",
+        "avxcfsn0.def",
+        "adcfra.def",
+        "avtchst0.def"
+      ]
+    }
+    
+    func exportVisitable() throws {
+        let visitableFileList = visitable.map { defFileName in
+            ImageExport(defFileName: defFileName, nameFromFrameIndex: { _ in defFileName })
+        }
+        
+        try generateTexture(
+            name: "visitable",
+            list: visitableFileList,
+            maxImageCountPerDefFile: 1
+        )
+    }
+}
