+++ Sources/Laka/Subcommands/Images/Textures/Textures.swift
@@ -1,0 +1,73 @@
+//
+//  File.swift
+//  
+//
+//  Created by Alexander Cyon on 2021-10-06.
+//
+
+import Foundation
+import ArgumentParser
+
+import Util
+import Malm
+import Guld
+
+// MARK: Textures
+extension Laka {
+    
+    /// A command to extract all textures, such as terrain, monsters, artifacts,
+    /// heroes and towns on the map etc. But not game menu UI, neither
+    /// in combat/battle creature sprites nor in town UI.
+    struct Textures: ParsableCommand, TextureGenerating {
+        
+        static var configuration = CommandConfiguration(
+            abstract: "Extract texture images such as terrain, artifacts, monsters etc. But not game menu UI."
+        )
+        
+        @OptionGroup var parentOptions: Options
+        
+    }
+}
+
+// MARK: Run
+extension Laka.Textures {
+    
+    mutating func run() throws {
+        print(
+            """
+            
+            🔮
+            About to extract textures from entries exported by the `Laka LOD` command.
+            Located at: \(inDataURL.path)
+            To folder: \(outImagesURL.path)
+            This will take about 30 seconds on a fast machine (Macbook Pro 2019 - Intel CPU)
+            ☕️
+            
+            """
+        )
+        try exportTerrain()
+        try exportTowns()
+        try exportMonsters()
+        try exportImpassableTerrain()
+        try exportPassableTerrain()
+        try exportVisitable()
+    }
+}
+
+
+// MARK: Computed props
+extension Laka.Textures {
+    
+    var verbose: Bool { parentOptions.printDebugInformation }
+    
+    var fileManager: FileManager { .default }
+    
+    var inDataURL: URL {
+        .init(fileURLWithPath: parentOptions.outputPath).appendingPathComponent("entries")
+    }
+    
+    var outImagesURL: URL {
+        .init(fileURLWithPath: parentOptions.outputPath).appendingPathComponent("images")
+    }
+}
+
